Vertices now defined wrt center of geometry
Vertex position definitions are now taken from the
full list of all vertices from all geometry. This
is done in an attempt to resolve the invalid
intersection problem.

All vertices are now defined with respect to the
center of the geometry. Previously, all geometry
was referenced from the vertex at the corner.